bd helper

+++ cartLibrary/src/main/java/com/android/tonyvu/sc/model/Saleable.java
@@ -1,12 +1,10 @@
 package com.android.tonyvu.sc.model;
 
-import java.math.BigDecimal;
-
 /**
  * Implements this interface for any product which can be added to shopping cart
  */
 public interface Saleable {
-    BigDecimal getPrice();
+    Float getPrice();
 
     String getName();
 }
